+++ src/IbanNet/Registry/IbanRegistry.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using IbanNet.Registry.Swift;
 
 namespace IbanNet.Registry
 {